controller: prevent claim storms around virtual PBs
When a port binding, which is a virtual parent of a virtual port
binding,
is bound to another chassis as additional, the virtual port also gets
bound to this new chassis.
The old chassis, which still is the requested chassis for the parent
port binding, claims the virtual port binding back.
This behaviour results in a claim storm for the virtual port binding.

This commit introduces a fix to prevent such behaviour. The virtual port
binding should only be bound to a chassis which is the requested
chassis, not additional chassis